+++ sandbox/order/models.py
@@ -1,8 +1,6 @@
 from oscar.apps.order.abstract_models import AbstractOrder, AbstractLine
-from oscar.core.loading import get_class
-
-CCHOrderMixin = get_class('oscarcch.mixins', 'CCHOrderMixin')
-CCHOrderLineMixin = get_class('oscarcch.mixins', 'CCHOrderLineMixin')
+from oscarcch.mixins import CCHOrderMixin
+from oscarcch.mixins import CCHOrderLineMixin
 
 
 class Order(CCHOrderMixin, AbstractOrder):