+++ src/oscarcch/tests/test_cch.py
@@ -1,10 +1,11 @@
+from unittest import mock
 from decimal import Decimal as D
 from freezegun import freeze_time
 from oscar.core.loading import get_model, get_class
 from oscar.test import factories
+from ..calculator import CCHTaxCalculator
 from .base import BaseTest
 from .base import p
-import mock
 import requests
 
 Basket = get_model('basket', 'Basket')
@@ -18,7 +19,6 @@ ConditionalOffer = get_model('offer', 'ConditionalOffer')
 
 USStrategy = get_class('partner.strategy', 'US')
 Applicator = get_class('offer.applicator', 'Applicator')
-CCHTaxCalculator = get_class('oscarcch.calculator', 'CCHTaxCalculator')
 
 
 class CCHTaxCalculatorTest(BaseTest):
@@ -187,8 +187,10 @@ class CCHTaxCalculatorTest(BaseTest):
 
         self.assertTrue(basket.is_tax_known)
         self.assertEqual(basket.total_excl_tax, D('5.00'))
-        self.assertEqual(basket.total_incl_tax, D('5.89'))
-        self.assertEqual(basket.total_tax, D('0.89'))
+        # self.assertEqual(basket.total_incl_tax, D('5.89'))
+        # self.assertEqual(basket.total_tax, D('0.89'))
+        self.assertEqual(basket.total_incl_tax, D('5.45'))
+        self.assertEqual(basket.total_tax, D('0.45'))
 
         purchase_info = basket.all_lines()[0].purchase_info
         self.assertEqual(purchase_info.price.excl_tax, D('10.00'))